Factored out ConfigureActions from Configure MenuButton
