+++ src/lib/y2partitioner/widgets/rescan_devices_button.rb
@@ -20,7 +20,7 @@
 require "yast"
 require "cwm/widget"
 require "y2partitioner/actions/rescan_devices"
-require "y2partitioner/widgets/execute_and_redraw"
+require "y2partitioner/execute_and_redraw"
 
 Yast.import "Popup"
 Yast.import "Mode"
